Tested jsm under sparse and dense.
